task_api_v1

+++ src/controllers/tasks.controller.js
@@ -1,10 +1,10 @@
 import { pool } from "../db.js";
+//TODO: Ver video de FAZT en el minuto 2:39:37 para continuar con el tutorial
 
+// Obtener todas las tareas
 export const getAllTasks = async (req, res, next) => {
-  
   try {
-    const result = await pool.query("SELECT * FROM tasks");
-    console.log(req.userID);
+    const result = await pool.query("SELECT * FROM tasks WHERE user_id = $1", [req.userID]);
     return res.status(200).json(result.rows);
   } catch (error) {
     console.error(error);
@@ -12,10 +12,11 @@ export const getAllTasks = async (req, res, next) => {
     next(error);
   }
 };
+// Obtener una tarea por su ID
 export const getTask = async (req, res, next) => {
   try {
     const { id } = req.params;
-    const result = await pool.query("SELECT * FROM tasks WHERE id = $1", [id]);
+    const result = await pool.query("SELECT * FROM tasks WHERE id = $1 AND user_id = $2", [id, req.userID]);
 
     if (result.rowCount === 0) {
       return res.status(404).json({ message: "Tarea no encontrada" });
@@ -30,10 +31,12 @@ export const getTask = async (req, res, next) => {
 // Crear una nueva tarea
 export const createTask = async (req, res, next) => {
   const { title, description } = req.body;
+  console.log(req.userID);
+  
   try {
     const result = await pool.query(
-      "INSERT INTO tasks (title, description) VALUES ($1, $2) RETURNING *", //Guardamos la tarea creada y devuelvemos esa misma tarea
-      [title, description]
+      "INSERT INTO tasks (title, description, user_id) VALUES ($1, $2, $3) RETURNING *", //Guardamos la tarea creada y devuelvemos esa misma tarea
+      [title, description, req.userID]
     );
     return res.status(201).json(result.rows[0]);
   } catch (error) {
@@ -48,14 +51,14 @@ export const createTask = async (req, res, next) => {
     next(error);
   }
 };
-
+// Actualizar una tarea existente
 export const updateTask = async (req, res, next) => {
   try {
     const { id } = req.params;
     const { title, description } = req.body;
     const result = await pool.query(
-      "UPDATE tasks SET title = $1, description= $2 WHERE id = $3 RETURNING *",
-      [title, description, id]
+      "UPDATE tasks SET title = $1, description= $2 WHERE id = $3 AND user_id = $4 RETURNING *",
+      [title, description, id, req.userID]
     );
     if (result.rowCount === 0) {
       return res.status(404).json({ message: "Tarea no encontrada" });
@@ -72,11 +75,11 @@ export const updateTask = async (req, res, next) => {
     next(error);
   }
 };
-
+// Eliminar una tarea
 export const deleteTask = async (req, res, next) => {
   try {
     const { id } = req.params;
-    const result = await pool.query("DELETE FROM tasks WHERE id = $1 ", [id]);
+    const result = await pool.query("DELETE FROM tasks WHERE id = $1 AND user_id = $2", [id, req.userID]);
     if (result.rowCount === 0) {
       return res.status(404).json({ message: "Tarea no encontrada" });
     }
